change the name Auth.java to AuthController.java
